Initial refactor of components
Move out from index every UI component to make it easier to maintain.
