+++ plugins/woocommerce-admin/client/payments-welcome/banner.tsx
@@ -1,7 +1,9 @@
 /**
  * External dependencies
  */
-import { Card, CardBody } from '@wordpress/components';
+import { Card, CardBody, CardHeader, Button } from '@wordpress/components';
+import { useState } from '@wordpress/element';
+import { recordEvent } from '@woocommerce/tracks';
 
 /**
  * Internal dependencies
@@ -9,10 +11,91 @@ import { Card, CardBody } from '@wordpress/components';
 import strings from './strings';
 import GiftIcon from './gift';
 import WCPayOfferIcon from './wcpay-offer.svg';
+import ExitSurveyModal from './exit-survey-modal';
+import PaymentMethods from './payment-methods';
 
-const Banner = () => {
+const LearnMore = () => {
+	const handleClick = () => {
+		recordEvent( 'wcpay_welcome_learn_more', {} );
+	};
 	return (
-		<Card size="large" className="account-page woocommerce-payments-banner">
+		<a
+			onClick={ handleClick }
+			href="https://woocommerce.com/payments/"
+			target="_blank"
+			rel="noreferrer"
+		>
+			{ strings.learnMore }
+		</a>
+	);
+};
+
+const Banner = ( {
+	isSubmitted,
+	handleSetup,
+}: {
+	isSubmitted: boolean;
+	handleSetup: () => void;
+} ) => {
+	const [ isNoThanksClicked, setNoThanksClicked ] = useState( false );
+
+	const [ isExitSurveyModalOpen, setExitSurveyModalOpen ] = useState( false );
+
+	const handleNoThanks = () => {
+		setNoThanksClicked( true );
+		setExitSurveyModalOpen( true );
+	};
+
+	return (
+		<Card className="connect-account__card">
+			<CardHeader>
+				<div>
+					<h1 className="banner-heading-copy">
+						{ strings.bannerHeading }
+					</h1>
+					<p className="wcpay-connect-account-page-terms-of-service">
+						{ strings.terms }
+					</p>
+				</div>
+				<div className="connect-account__action">
+					<Button
+						isSecondary
+						isBusy={ isNoThanksClicked && isExitSurveyModalOpen }
+						disabled={ isNoThanksClicked && isExitSurveyModalOpen }
+						onClick={ handleNoThanks }
+						className="btn-nothanks"
+					>
+						{ strings.nothanks }
+					</Button>
+					<Button
+						isPrimary
+						isBusy={ isSubmitted }
+						disabled={ isSubmitted }
+						onClick={ handleSetup }
+						className="btn-install"
+					>
+						{ strings.button }
+					</Button>
+					{ isExitSurveyModalOpen && (
+						<ExitSurveyModal
+							setExitSurveyModalOpen={ setExitSurveyModalOpen }
+						/>
+					) }
+				</div>
+			</CardHeader>
+			<CardBody>
+				<div className="content">
+					<p className="onboarding-description">
+						{ strings.onboarding.description } <LearnMore />
+					</p>
+
+					<p className="accepted-payment-methods">
+						{ strings.paymentMethodsHeading }
+					</p>
+
+					<PaymentMethods />
+				</div>
+			</CardBody>
 			<CardBody>
 				<div className="limited-time-offer">
 					<div className="offer-header">
